Add files via upload
Embark on a transformative journey as we develop a sophisticated chatbot in Dialogflow tailored for a food delivery system. This end-to-end project will comprehensively cover Dialogflow basics, guide you through the process of building a backend in Python using FastAPI, and showcase seamless interactions with a MySQL database. Throughout the project, we will explore Dialogflow fundamentals such as intents, entities, contexts, providing you with a solid foundation for creating intelligent conversational interfaces.

Acknowledging the valuable contribution of Codebasics, we will showcase the integration of FastAPI, a cutting-edge web framework, highlighting its prowess in facilitating communication between Dialogflow and the backend. The inclusion of a MySQL database adds a layer of data management, enhancing the overall functionality and efficiency of the food delivery system.

By the conclusion of this project, you will have gained a comprehensive understanding of building an intelligent chatbot for a food delivery system. Whether you're a beginner exploring conversational interfaces or an experienced developer looking to enhance your skill set, this project is designed to empower you with the skills needed to create an efficient chatbot. Join us on this transformative journey as we navigate the realms of Dialogflow, FastAPI, and MySQL, creating a seamless and intuitive solution for the food delivery industry.